being overly ambitious

+++ src/main.py
@@ -4,8 +4,9 @@
 
 #Run our 3 experiments on all datasets
 
+from typing import Callable
 from cleanup import cleanup
-from load_data_time_series.HAR.e4_wristband_Nov2019 import e4_load_dataset
+from load_data_time_series.HAR.e4_wristband_Nov2019.e4_load_dataset import e4_load_dataset
 from load_data_time_series.HAR.UniMiB_SHAR.unimib_shar_adl_load_dataset import unimib_load_dataset
 from load_data_time_series.HAR.UCI_HAR.uci_har_load_dataset import uci_har_load_dataset
 from utils.sh_loader import sh_loco_load_dataset
@@ -14,6 +15,7 @@ from experiment2 import exp_2
 from experiment3 import exp_3
 import numpy as np
 import pandas as pd
+import os
 
 CLEANUP = False
 
@@ -32,15 +34,21 @@ def channel_swap(X : np.ndarray) -> np.ndarray:
     assert X.ndim == 3, "Data must be 3-dimensional to channel swap"
     return np.reshape(X, (X.shape[0], X.shape[2], X.shape[1]))
 
-def run_and_write(exp: function, X_train: np.ndarray, y_train: np.ndarray, X_test: np.ndarray, y_test: np.ndarray, set: str, filename: str) -> None:
+def run_and_write(exp: Callable, X_train: np.ndarray, y_train: np.ndarray, X_test: np.ndarray, y_test: np.ndarray, set: str, filename: str) -> None:
     """
     Open the resutlts file
     Run one experiment on one dataset
     Add the new results to the file
     """
-    results = exp(X_train, y_train, X_test, y_test)
+    results = exp(X_train, y_train, X_test, y_test, set)
     results = pd.DataFrame.from_dict(results)
-    results.to_csv(filename)
+    if not os.path.exists(filename):      
+        results.to_csv(filename)
+    else:
+        file_frame = pd.read_csv(filename)
+        file_frame.append(results)
+        file_frame.to_csv(filename)
+
 
 
 if __name__ == '__main__':
